+++ scripts/invocation-repelling-blast.js
@@ -1,0 +1,82 @@
+const version = "0.1.0";
+const optionName = "Repelling Blast";
+
+try {
+	if (args[0].macroPass === "DamageBonus") {
+		const lastArg = args[args.length - 1];
+		const attacker = canvas.tokens.get(args[0].tokenId);
+		let ttoken = canvas.tokens.get(args[0].hitTargets[0].object.id);
+		/*
+		let workflow = MidiQOL.Workflow.getWorkflow(args[0].uuid);
+		let actor = workflow?.actor;
+		let target = args[0].hitTargets[0];
+		let tactor = target?.actor;
+		let ttoken = canvas.tokens.get(args[0].hitTargets[0].object.id);
+		let pusher = canvas.tokens.get(args[0].tokenId);
+		*/
+		
+		// make sure it's an allowed attack
+		if (lastArg.item.name !== "Eldritch Blast") {
+			console.log(`${optionName}: not Eldricth Blast`);
+			return {};
+		}
+		
+		// ask if they want to use the option
+		let dialog = new Promise((resolve, reject) => {
+			new Dialog({
+				// localize this text
+				title: `Eldritch Invocation: ${optionName}`,
+				content: `<p>Use ${optionName} to push ${ttoken.name} 10 feet?</p>`,
+				buttons: {
+					one: {
+						icon: '<p> </p><img src = "icons/magic/sonic/projectile-shock-wave-blue.webp" width="30" height="30"></>',
+						label: "<p>Yes</p>",
+						callback: () => resolve(true)
+					},
+					two: {
+						icon: '<p> </p><img src = "icons/skills/melee/weapons-crossed-swords-yellow.webp" width="30" height="30"></>',
+						label: "<p>No</p>",
+						callback: () => { resolve(false) }
+					}
+				},
+				default: "two"
+			}).render(true);
+		});
+		
+		let useManeuver = await dialog;
+		if (useManeuver) {
+			await pushTarget(attacker, ttoken);
+		}
+		
+	}
+
+	return {};
+	
+} catch (err) {
+    console.error(`${optionName} - ${version}`, err);
+}
+
+async function pushTarget(sourceToken, targetToken) {
+	let newCenter = getAllowedPushLocation(sourceToken, targetToken, 2);
+	if(!newCenter) {
+		return ui.notifications.error(`${optionName} - no room to push ${targetToken.name}`);
+	}
+	const tobj = targetToken.document.object;
+	newCenter = canvas.grid.getSnappedPosition(newCenter.x - tobj.w / 2, newCenter.y - tobj.h / 2, 1);
+	const mutationData = { token: {x: newCenter.x, y: newCenter.y}};
+	await warpgate.mutate(targetToken.document, mutationData, {}, {permanent: true});
+}
+
+function getAllowedPushLocation(sourceToken, targetToken, maxSquares) {
+	for (let i = maxSquares; i > 0; i--) {
+		const knockbackPixels = i * canvas.grid.size;
+		const ray = new Ray(sourceToken.center, targetToken.center);
+		const newCenter = ray.project((ray.distance + knockbackPixels)/ray.distance);
+		const isAllowedLocation = canvas.sight.testVisibility({x: newCenter.x, y: newCenter.y}, {object: targetToken.Object});
+		if(isAllowedLocation) {
+			return newCenter;
+		}
+	}
+	
+	return null;
+}